Generate a sheets.json
This JSON file contains one two-dimensional table, where each row
correspons to a video. The first column contains the video identifier,
and the colums after the first contain the names of contributors.

The idea is, that is JSON file can later contain more tables, that can
be exported to Google Sheets.

It looks like this:

```
{
  "contributors_per_video": [
    [
      "Video",
      "Contributors"
    ],
    [
      "2015/binary-counting",
      "Grant Sanderson",
      "Omer Tuchfeld",
      "bot224032",
      "luiz12apn"
    ],
    [
      "2015/eulers-characteristic-formula",
      "Grant Sanderson",
      "realcalal",
      "Omer Tuchfeld"
    ],
    [
      "2015/eulers-formula-old",
      "Grant Sanderson"
    ],
    // ...
  ]
}
```